requirements.txt
asgiref==3.3.1
boto==2.49.0
boto3==1.17.36
botocore==1.20.36
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-storages==1.11.1
djangorestframework==3.12.3
jmespath==0.10.0
Pillow==8.1.2
psycopg2==2.8.6
PyJWT==2.0.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
